Implement ender crystals (#530)
* Add Ender Crystal entity

* Make Ender Crystal craftable and placeable

* Adds storage and basic api stuff

* Ender Crystals explode upon interaction

Related to #504

+++ src/main/java/net/glowstone/util/Position.java
@@ -138,7 +138,7 @@ public final class Position {
      * @return the serialized position value
      */
     public static long getPosition(BlockVector vector) {
-        return ((vector.getBlockX() & 0x3FFFFFF) << 38) | ((vector.getBlockY() & 0xFFF) << 26) | (vector.getBlockZ() & 0x3FFFFFF);
+        return (((long) vector.getBlockX() & 0x3FFFFFF) << 38) | (((long) vector.getBlockY() & 0xFFF) << 26) | ((long) vector.getBlockZ() & 0x3FFFFFF);
     }
 
     /**
